\name{overview}
\alias{overview}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Overview function
}
\description{
This functions can be used to show the graphics generated by the functions dynRB_Pn,dynRB_Vn and dynRB_VPa.
}
\usage{
overview(r, row_col = c(3, 3))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{r}{
Output of the function dynRB_Pn,dynRB_Vn or dynRB_VPa.
}
  \item{row_col}{
  Number of rows and columns of the figures. Default: \code{row_col = c(3, 3)}
}
}
\author{
Manuela Schreyer \email{manuelalarissa.schreyer@sbg.ac.at},\cr 
Wolfgang Trutschnig \email{Wolfgang.Trutschnig@sbg.ac.at},\cr 
Robert R. Junker \email{Robert.Junker@sbg.ac.at} (corresponding author),\cr 
Jonas Kuppler \email{Jonas.Kuppler@uni-duesseldorf.de},\cr 
Arne Bathke \email{Arne.Bathke@sbg.ac.at}
}

\examples{
# example for the function dynRB_Pn
# for reliable results use steps = 201
data(finch2)
r<-dynRB_Pn(finch2, steps = 101)
overview(r)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
